backtesting/requirements.txt
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.28
requests==2.32.3
PyYAML==6.0.2
pytest==8.3.2
pandas-market-calendars==4.4.0
matplotlib==3.9.2
Jinja2==3.1.4
python-dateutil==2.9.0.post0
ruff==0.6.9
pyarrow==17.0.0
